[Proposal]: Alternative way of timestampoffset handling on the PS4
This is a very minor reimplementation of #1432 where the code updating
the tfdt is now in the `SourceBufferInterface` implementation instead of
in its own compat file.

This has the disadvantage of making the `SourceBufferInterface` class a
little more complex by adding a lot of code only useful for the PlayStation
4 for now in it.

Its advantage however is that by integrating the work-around in this
tighter way, we better ensure the initialization segment's timescale
considered is up to date and resetted at the right time.
